+++ Recurly/Resources/SubscriptionAddOn.cs
@@ -75,7 +75,7 @@ namespace Recurly.Resources
 
         /// <value>This is priced in the subscription's currency.</value>
         [JsonProperty("unit_amount")]
-        public float? UnitAmount { get; set; }
+        public decimal? UnitAmount { get; set; }
 
         /// <value>Updated at</value>
         [JsonProperty("updated_at")]
@@ -83,7 +83,7 @@ namespace Recurly.Resources
 
         /// <value>The percentage taken of the monetary amount of usage tracked. This can be up to 4 decimal places. A value between 0.0 and 100.0. Required if add_on_type is usage and usage_type is percentage.</value>
         [JsonProperty("usage_percentage")]
-        public float? UsagePercentage { get; set; }
+        public decimal? UsagePercentage { get; set; }
 
     }
 }
