+++ Recurly/Resources/Transaction.cs
@@ -21,7 +21,7 @@ namespace Recurly.Resources
 
         /// <value>Total transaction amount sent to the payment gateway.</value>
         [JsonProperty("amount")]
-        public float? Amount { get; set; }
+        public decimal? Amount { get; set; }
 
         /// <value>When processed, result from checking the overall AVS on the transaction.</value>
         [JsonProperty("avs_check")]
@@ -80,7 +80,7 @@ namespace Recurly.Resources
 
         /// <value>Time, in seconds, for gateway to process the transaction.</value>
         [JsonProperty("gateway_response_time")]
-        public float? GatewayResponseTime { get; set; }
+        public decimal? GatewayResponseTime { get; set; }
 
         /// <value>The values in this field will vary from gateway to gateway.</value>
         [JsonProperty("gateway_response_values")]
